Fixed build warning '=': conversion from 'Uint32' to 'Uint16', possible loss of data

+++ src/joystick/hidapi/SDL_hidapi_ps5.c
@@ -1315,7 +1315,7 @@ static void HIDAPI_DriverPS5_HandleStatePacketCommon(SDL_Joystick *joystick, SDL
 
         if (ctx->use_alternate_report) {
             /* 16-bit timestamp */
-            Uint16 delta;
+            Uint32 delta;
             Uint16 tick = LOAD16(packet->rgucSensorTimestamp[0],
                                  packet->rgucSensorTimestamp[1]);
             if (ctx->last_tick < tick) {
